Added enums to automation interface. Implemented support for transparent block borders.

+++ src/Automation/ProMoBoundaryBlocksAuto.h
@@ -51,7 +51,7 @@ protected:
 	//{{AFX_DISPATCH(CProMoBoundaryBlocksAuto)
 	afx_msg VARIANT GetIDs();
 	afx_msg void SetIDs(const VARIANT FAR& newValue);
-	afx_msg BOOL Add(LPDISPATCH Item, short Attachment);
+	afx_msg BOOL Add(LPDISPATCH Item, long Attachment);
 	afx_msg short Count();
 	afx_msg BOOL Remove(const VARIANT FAR& Item);
 	afx_msg LPDISPATCH GetItem(const VARIANT FAR& Item);
